+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 - added options to download sources and Javadoc via IDEA Gradle plugin
 - added option to publish the default access transformer
+- added option to define custom file targets for process resources task
 
 ## [2.2.0] - 2026-06-18